refactor(codegen): move MIR utilities into MIR
