+++ crates/codegen/src/transform/dce.rs
@@ -4,9 +4,8 @@
 
 use crate::{
     analysis::CfgInfo,
-    mir::{BlockId, Function, InstId, Terminator, Value, ValueId},
+    mir::{BlockId, Function, InstId, Terminator, Value, ValueId, utils::repair_reachability_phis},
     pass::FunctionPass,
-    utils::repair_reachability_phis,
 };
 use solar_data_structures::map::{FxHashMap, FxHashSet};
