+++ crates/codegen/src/transform/loop_opt.rs
@@ -12,9 +12,11 @@
 
 use crate::{
     analysis::{AffineExpr, Loop, LoopAnalyzer, ScalarEvolution},
-    mir::{BlockId, Function, InstId, InstKind, StorageAlias, Terminator, Value, ValueId},
+    mir::{
+        BlockId, Function, InstId, InstKind, StorageAlias, Terminator, Value, ValueId,
+        utils as mir_utils,
+    },
     pass::FunctionPass,
-    utils::mir as mir_utils,
 };
 use alloy_primitives::U256;
 use solar_data_structures::map::FxHashSet;
