--- crates/codegen/src/utils/cfg.rs
@@ -1,110 +1,0 @@
-//! CFG mutation utilities shared by MIR analyses and transformation passes.
-
-use crate::mir::{BasicBlock, BlockId, Function, InstKind, Terminator};
-use smallvec::smallvec;
-
-/// Splits the CFG edge from `pred` to `succ` by inserting a fresh block that
-/// contains no instructions and jumps straight to `succ`. Returns the new
-/// block.
-///
-/// A terminator that targets `succ` several times — a branch with both arms to
-/// `succ`, or a switch with several cases plus the default — is one logical
-/// edge: every occurrence is retargeted to the same new block. Phi incoming
-/// lists are keyed per predecessor block, so two distinct split blocks for one
-/// predecessor would force conflicting phi entries.
-///
-/// Self-loops (`pred == succ`) are supported: the new block takes over the
-/// backedge and `succ`'s phis are rekeyed from `pred` to the new block.
-///
-/// `succ` cannot be the entry block, which has no predecessors by definition.
-pub(crate) fn split_edge(func: &mut Function, pred: BlockId, succ: BlockId) -> BlockId {
-    debug_assert_ne!(succ, func.entry_block, "the entry block has no predecessor edges to split");
-
-    let new_block = func.blocks.push(BasicBlock {
-        instructions: Vec::new(),
-        terminator: Some(Terminator::Jump(succ)),
-        predecessors: smallvec![pred],
-    });
-
-    // Retarget every occurrence of `succ` among `pred`'s successors.
-    let mut retargeted = false;
-    let mut retarget = |block: &mut BlockId| {
-        if *block == succ {
-            *block = new_block;
-            retargeted = true;
-        }
-    };
-    match func.blocks[pred].terminator.as_mut().expect("predecessor must have a terminator") {
-        Terminator::Jump(target) => retarget(target),
-        Terminator::Branch { then_block, else_block, .. } => {
-            retarget(then_block);
-            retarget(else_block);
-        }
-        Terminator::Switch { default, cases, .. } => {
-            retarget(default);
-            for (_, case_block) in cases {
-                retarget(case_block);
-            }
-        }
-        term => panic!("terminator `{}` has no successor edges to split", term.mnemonic()),
-    }
-    assert!(retargeted, "bb{} does not branch to bb{}", pred.index(), succ.index());
-
-    // Replace `pred` with the new block in `succ`'s predecessor list. A
-    // multi-occurrence edge may have been recorded once per occurrence;
-    // they all collapse into the single split block.
-    let predecessors = &mut func.blocks[succ].predecessors;
-    let first = predecessors
-        .iter()
-        .position(|&block| block == pred)
-        .expect("successor must list pred as a predecessor");
-    predecessors[first] = new_block;
-    predecessors.retain(|&mut block| block != pred);
-
-    // Rekey `succ`'s phi incoming entries from `pred` to the new block.
-    for &inst_id in &func.blocks[succ].instructions {
-        if let InstKind::Phi(incoming) = &mut func.instructions[inst_id].kind {
-            for (block, _) in incoming.iter_mut() {
-                if *block == pred {
-                    *block = new_block;
-                }
-            }
-        }
-    }
-
-    new_block
-}
-
-/// Rebuilds CFG edge lists from terminators and drops phi inputs from blocks
-/// that are no longer predecessors. Returns true if any phi input was dropped.
-pub(crate) fn repair_reachability_phis(func: &mut Function) -> bool {
-    let mut edges = Vec::new();
-    for (block, bb) in func.blocks.iter_enumerated() {
-        if let Some(term) = &bb.terminator {
-            edges.push((block, term.successors()));
-        }
-    }
-
-    for block in func.blocks.iter_mut() {
-        block.predecessors.clear();
-    }
-
-    for (block, successors) in edges {
-        for succ in successors {
-            func.blocks[succ].predecessors.push(block);
-        }
-    }
-
-    let mut changed = false;
-    for block_id in func.blocks.indices() {
-        let predecessors = func.blocks[block_id].predecessors.clone();
-        for &inst_id in &func.blocks[block_id].instructions {
-            if let InstKind::Phi(incoming) = &mut func.instructions[inst_id].kind {
-                let len_before = incoming.len();
-                incoming.retain(|(pred, _)| predecessors.contains(pred));
-                changed |= incoming.len() != len_before;
-            }
-        }
-    }
-    changed
-}
